copiado projeto do Massaki a fim de compreender as falhas
