+++ src/actions/index.js
@@ -1,1 +1,62 @@
 // Coloque aqui suas actions
+import currenciesAPI from '../services';
+// Coloque aqui suas actions
+export const EMAIL = 'EMAIL';
+export const REQUEST_START = 'REQUEST_START';
+export const REQUEST_SUCCESS = 'REQUEST_SUCCESS';
+export const REQUEST_FAIL = 'REQUEST_FAIL';
+export const SAVE_EXPENSE = 'SAVE_EXPENSE';
+export const DELETE_EXPENSE = 'DELETE_EXPENSE';
+export const EDIT_EXPENSE_START = 'EDIT_EXPENSE_START';
+export const EDIT_EXPENSE_END = 'EDIT_EXPENSE_END';
+
+export const emailChange = (payload) => ({
+  type: EMAIL,
+  payload,
+});
+
+const requestCurrenciesTry = () => ({
+  type: REQUEST_START,
+});
+
+const requestCurrenciesSuccess = (currencies) => ({
+  type: REQUEST_SUCCESS,
+  currencies,
+});
+
+const requestCurrenciesFail = (error) => ({
+  type: REQUEST_FAIL,
+  error,
+});
+
+export const fetchCurrencies = () => async (dispatch) => {
+  try {
+    dispatch(requestCurrenciesTry());
+
+    const currencies = await currenciesAPI();
+
+    dispatch(requestCurrenciesSuccess(currencies));
+  } catch (error) {
+    dispatch(requestCurrenciesFail(error));
+  }
+};
+
+export const saveExpense = (expenses) => ({
+  type: SAVE_EXPENSE,
+  expenses,
+});
+
+export const deleteExpense = (expense) => ({
+  type: DELETE_EXPENSE,
+  expense,
+});
+
+export const editExpense = (expense) => ({
+  type: EDIT_EXPENSE_START,
+  expense,
+});
+
+export const endExpenseEdit = (expense) => ({
+  type: EDIT_EXPENSE_END,
+  expense,
+});
