+++ src/components/Header.js
@@ -1,0 +1,53 @@
+import React from 'react';
+import PropTypes from 'prop-types';
+import { connect } from 'react-redux';
+
+class Header extends React.Component {
+  constructor() {
+    super();
+
+    this.getTotalExpenses = this.getTotalExpenses.bind(this);
+  }
+
+  getTotalExpenses() {
+    const { expenses } = this.props;
+    const totalExpenses = expenses.reduce((total, each) => {
+      const { value, currency, exchangeRates } = each;
+      const rate = parseFloat(exchangeRates[currency].ask);
+      return total + parseFloat(value) * rate;
+    }, 0);
+    return totalExpenses.toFixed(2);
+  }
+
+  render() {
+    const { email } = this.props;
+    return (
+      <header className="wallet-header">
+        <h1>Trybe Wallet</h1>
+        <p data-testid="email-field" className="wallet-email">{`E-mail: ${email}`}</p>
+        <div className="wallet-total-value">
+          <span data-testid="total-field">
+            {`Despesa Total: R$ ${this.getTotalExpenses()} `}
+          </span>
+          <span data-testid="header-currency-field">BRL</span>
+        </div>
+      </header>
+    );
+  }
+}
+
+const mapStateToProps = (state) => ({
+  email: state.user.email,
+  expenses: state.wallet.expenses,
+});
+
+export default connect(mapStateToProps)(Header);
+
+Header.propTypes = {
+  email: PropTypes.string.isRequired,
+  expenses: PropTypes.arrayOf(PropTypes.object),
+};
+
+Header.defaultProps = {
+  expenses: [],
+};
