+++ src/components/index.js
@@ -1,0 +1,4 @@
+export { default as Header } from './Header';
+export { default as NewExpenseForm } from './NewExpenseForm';
+export { default as EditExpenseForm } from './EditExpenseForm';
+export { default as ExpensesTable } from './ExpensesTable';
