+++ src/reducers/wallet.js
@@ -1,35 +1,41 @@
 // Esse reducer será responsável por tratar o todas as informações relacionadas as despesas
 
-import * as walletAction from '../actions/wallet';
+import {
+  REQUEST_START,
+  REQUEST_SUCCESS,
+  REQUEST_FAIL,
+  SAVE_EXPENSE,
+  DELETE_EXPENSE,
+  EDIT_EXPENSE_START,
+  EDIT_EXPENSE_END,
+} from '../actions';
 
-const initialStateWallet = {
+const INITIAL_STATE_WALLET = {
   currencies: [],
   expenses: [],
-  loading: false,
 };
 
-const wallet = (state = initialStateWallet, action) => {
+export default function wallet(state = INITIAL_STATE_WALLET, action) {
   switch (action.type) {
-  case walletAction.REQUEST_TO_API:
-    return { ...state, loading: true };
-
-  case walletAction.REQUEST_FAIL:
-    return { ...state, loading: false, error: action.error };
-
-  case walletAction.REQUEST_SUCESS:
+  case REQUEST_START:
     return {
       ...state,
-      loading: false,
-      currencies: [...Object.keys(action.currencies)
-        .filter((noUSTD) => noUSTD !== 'USDT')],
+      isFetching: true,
     };
-
-  case walletAction.ADD_EXPENSE:
+  case REQUEST_SUCCESS:
     return {
-      ...state, expenses: [...state.expenses, action.expenses],
+      ...state,
+      isFetching: false,
+      currencies: [...Object.keys(action.currencies)],
     };
-
-  case walletAction.DELETE_EXPENSE:
+  case REQUEST_FAIL:
+    return { ...state, isFetching: false, error: action.error };
+  case SAVE_EXPENSE:
+    return {
+      ...state,
+      expenses: [...state.expenses, action.expenses],
+    };
+  case DELETE_EXPENSE:
     return {
       ...state,
       expenses: [
@@ -37,13 +43,13 @@ const wallet = (state = initialStateWallet, action) => {
       ],
       isEditing: false,
     };
-  case walletAction.INIT_EDIT_EXPENSE:
+  case EDIT_EXPENSE_START:
     return {
       ...state,
       isEditing: true,
       expenseId: action.expense.id,
     };
-  case walletAction.FINISH_EDIT_EXPENSE:
+  case EDIT_EXPENSE_END:
     return {
       ...state,
       expenses: state.expenses.map((item) => {
@@ -52,10 +58,67 @@ const wallet = (state = initialStateWallet, action) => {
       }),
       isEditing: false,
     };
-
   default:
-    return initialStateWallet;
+    return state;
   }
-};
+}
 
-export default wallet;
+// import * as walletAction from '../actions/wallet';
+
+// const initialStateWallet = {
+//   currencies: [],
+//   expenses: [],
+//   loading: false,
+// };
+
+// const wallet = (state = initialStateWallet, action) => {
+//   switch (action.type) {
+//   case walletAction.REQUEST_TO_API:
+//     return { ...state, loading: true };
+
+//   case walletAction.REQUEST_FAIL:
+//     return { ...state, loading: false, error: action.error };
+
+//   case walletAction.REQUEST_SUCESS:
+//     return {
+//       ...state,
+//       loading: false,
+//       currencies: [...Object.keys(action.currencies)
+//         .filter((noUSTD) => noUSTD !== 'USDT')],
+//     };
+
+//   case walletAction.ADD_EXPENSE:
+//     return {
+//       ...state, expenses: [...state.expenses, action.expenses],
+//     };
+
+//   case walletAction.DELETE_EXPENSE:
+//     return {
+//       ...state,
+//       expenses: [
+//         ...state.expenses.filter((expense) => expense.id !== action.expense.id),
+//       ],
+//       isEditing: false,
+//     };
+//   case walletAction.INIT_EDIT_EXPENSE:
+//     return {
+//       ...state,
+//       isEditing: true,
+//       expenseId: action.expense.id,
+//     };
+//   case walletAction.FINISH_EDIT_EXPENSE:
+//     return {
+//       ...state,
+//       expenses: state.expenses.map((item) => {
+//         if (item.id === action.expense.id) return { ...item, ...action.expense };
+//         return item;
+//       }),
+//       isEditing: false,
+//     };
+
+//   default:
+//     return initialStateWallet;
+//   }
+// };
+
+// export default wallet;
